switch to decimal

+++ src/Modules/SimpleCommandsModule.cs
@@ -10,9 +10,9 @@ namespace Better_BelieveIt_Bot.Modules {
 
         [SlashCommand("math_add", "Add two integers together.")]
         public async Task MathAdd(
-            [Summary(description: "The first number to add")] int a,
-            [Summary(description: "The second number to add")] int b) {
-            long c = a + b;
+            [Summary(description: "The first number to add")] decimal a,
+            [Summary(description: "The second number to add")] decimal b) {
+            decimal c = a + b;
             await RespondAsync($"The sum of [{a}] and [{b}] is [{c}]");
         }
     }
